Trapped zone model (first try...)
Input changes :
      attributes are now set in a loop
      units of each input parameter must be stored
      in the new dictionnary d_inputs{}

Time integration :
     it is now performed by the Time_integration(Y0,time) method
     (wrapper to call the Scipy odeint routine)

     just comment the line to go back to the previous version
     using the Eval_Omega(time) method

The ODE system to integrate in time is defined in
the Build_RHS() method. Each equation RHS is computed
by calling a method named under the form *_dot() and based
on the model of the existing Omega_dot(Omega, T) method.

To do that, the dipole and propeller luminosities have been
recomputed in the methods
- Luminosity_dipole(Omega,T)
- Luminosity_propeller(Omega,T)

Remarks:
- new outputs are not implemented yet
- the signature of the methods is not consistent throughout the code;
one must be careful to the order of the variables !