Added the trapped and free x-ray luminositied outputs.
From Sun et al. 2017 we added the xray luminosity from the dissipation
of the NS wind (dipole spindown and propeller with individual
efficiencies) and the xray luminosity from the trapped zone (residual
emission from the wind + black-body component).

Added methods to evaluate the new luminosity components and the BB-like
temperature.

Added a few new constants (Planck, Boltzmann and radiation density).

Added an extra plot for the new luminosities.

The Build_RHS method returns an homogeneous array, but currently a
slicing [:,0] is required to get the correct shape for the output.