Handling different EOS
Use a dictionary to store the different EOS, which
can be then passed to the GRB() method like other
parameters:

GRB(**inputs,**EOS['GM1'])

Rem: this avoids potential conflicts with duplicate values
in the two dictionaries.

Replaced 'None' as a default value by '-np.inf'
to simplify printing format (try to keep the same
type for variables) ; removed the 'None' default value
in the methods where it was never used.

Commentedsome old method calls in __init__().

Updated the Info() methods with the new output parameters
(and renamed some variables).